Eliminate ListItemSelectionTrait and fix token translation.
